cloud/proxy/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
aiohttp==3.8.1
PyJWT==2.3.0
python-multipart==0.0.5 